+++ app/Models/RegisterProject.php
@@ -13,7 +13,7 @@ class RegisterProject extends Model
      *
      * @var array
      */
-    protected $fillable = ['user_id', 'project_id'];
+    protected $fillable = ['user_id', 'project_id','payment_id','payment_request_id','price'];
 
     protected static function addFavouriteProject(Request $request){
     	$userId = $request->get('user_id');
@@ -24,10 +24,13 @@ class RegisterProject extends Model
     			$registeredProject->save();
                 return 'true';
     		} else {
-                $registeredProject->delete();
-                return 'false';
+                if(empty($registeredProject->payment_id) && empty($registeredProject->payment_request_id) && empty($registeredProject->price)){
+                    $registeredProject->delete();
+                    return 'false';
+                }
             }
     	}
+        return 'true';
     }
 
     protected static function getRegisteredProjectsByUserId($userId){
@@ -53,6 +56,10 @@ class RegisterProject extends Model
         return static::where('user_id', $userId)->get();
     }
 
+    protected static function getRegisteredVkitProjectByUserIdByProjectId($userId,$projectId){
+        return static::where('user_id', $userId)->where('project_id',$projectId)->first();
+    }
+
     protected static function deleteRegisteredVkitProjectsByUserId($userId){
         $projects = static::where('user_id', $userId)->get();
         if(is_object($projects) && false == $projects->isEmpty()){
@@ -61,5 +68,33 @@ class RegisterProject extends Model
             }
         }
         return;
+    }
+
+    protected static function addPurchasedProject($paymentArray){
+        $purchasedProject = static::firstOrNew(['user_id' => $paymentArray['user_id'], 'project_id' => $paymentArray['project_id']]);
+        if(is_object($purchasedProject) && empty($purchasedProject->id)){
+            $purchasedProject->payment_id = $paymentArray['payment_id'];
+            $purchasedProject->payment_request_id = $paymentArray['payment_request_id'];
+            $purchasedProject->price = $paymentArray['price'];
+            $purchasedProject->save();
+        } else {
+            $purchasedProject->payment_id = $paymentArray['payment_id'];
+            $purchasedProject->payment_request_id = $paymentArray['payment_request_id'];
+            $purchasedProject->price = $paymentArray['price'];
+            $purchasedProject->save();
+        }
+        return $purchasedProject;
+    }
+
+    protected static function getPurchasedProjectsByUserIdForPayments($userId){
+        return static::join('vkit_projects','vkit_projects.id','=','register_projects.project_id')
+            ->whereNotNull('register_projects.payment_id')
+            ->whereNotNull('register_projects.payment_request_id')
+            ->where('register_projects.price', '>', 0)
+            ->whereNotNull('register_projects.payment_id')
+            ->whereNotNull('register_projects.payment_request_id')
+            ->where('register_projects.user_id', $userId)
+            ->select('register_projects.id','register_projects.updated_at','register_projects.price','vkit_projects.name')
+            ->groupBy('register_projects.id')->get();
     }
 }
